+++ CHANGELOG.md
@@ -8,13 +8,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- Added Angular integration.
-- Added patch-package to remove jsxDev from Angular build.
+- Added Testing Library integration testing for React and Vue.
 - Added Vitest unit testing.
+- Added patch-package to remove jsxDev from Angular build.
+- Added Angular integration.
 
 ### Fixed
-- Fixed Bun build and output.
 - Fixed Deno build and output.
+- Fixed Bun build and output.
 
 ## [0.3.0] - 2025-12-24